[SofaOpenglVisual] UPDATE: enable ogl label position as fraction of screen
(cherry picked from commit 3b94371d9d21f31b75dbd674f271e3a3f3f6fe41)

+++ modules/SofaOpenglVisual/OglLabel.h
@@ -62,8 +62,8 @@ public:
     Data<std::string> prefix;
     Data<std::string> label;
     Data<std::string> suffix;
-    Data<unsigned int> x;
-    Data<unsigned int> y;
+    Data<float> x;
+    Data<float> y;
     Data<unsigned int> fontsize;
     Data<std::string> color;
     Data<unsigned int> updateLabelEveryNbSteps;
